server/pyproject.toml
[project]
name = "bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.5",
    "aiokafka==0.12.0",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "async-timeout==5.0.1",
    "bcrypt==4.2.1",
    "billiard==4.2.1",
    "boto3==1.36.11",
    "botocore==1.36.11",
    "celery==5.4.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "click==8.1.7",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.115.6",
    "fastapi-cli==0.0.6",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.4",
    "jmespath==1.0.1",
    "kafka-python>=2.0.6",
    "kombu==5.4.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "motor==3.6.0",
    "packaging==24.2",
    "passlib==1.7.4",
    "pika>=1.3.2",
    "pillow>=11.1.0",
    "pillow-avif-plugin>=1.4.6",
    "prompt-toolkit==3.0.50",
    "pycparser==2.22",
    "pydantic==2.10.3",
    "pydantic-core==2.27.1",
    "pydantic-settings==2.7.1",
    "pygments==2.18.0",
    "pyjwt==2.10.1",
    "pymongo==4.9.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.19",
    "pyyaml==6.0.2",
    "redis==5.2.1",
    "rich==13.9.4",
    "rich-toolkit==0.12.0",
    "s3transfer==0.11.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "uvicorn==0.32.1",
    "uvloop==0.21.0",
    "vine==5.1.0",
    "watchfiles==1.0.3",
    "wcwidth==0.2.13",
    "websockets==14.1",
]
